backend/damage_model.py
from pathlib import Path
from PIL import Image
import numpy as np
from typing import Dict, Any

class DamageAnalyzer:
    def __init__(self):
        # Initialize model parameters
        self.damage_types = [
            "scratch", "dent", "crack", "broken_part"
        ]
        self.severity_levels = [
            "minor", "moderate", "severe"
        ]
        
    def preprocess_image(self, image_path: str) -> np.ndarray:
        """Preprocess image for analysis"""
        img = Image.open(image_path)
        img = img.resize((224, 224))  # Standard input size
        return np.array(img)
    
    def analyze(self, image_array: np.ndarray) -> Dict[str, Any]:
        """
        Analyze vehicle damage from preprocessed image
        Note: This is a placeholder implementation
        """
        # Simulate damage analysis
        # In a real implementation, this would use a trained ML model
        return {
            "damage_type": "dent",
            "severity": "moderate",
            "location": "front_bumper",
            "confidence": 0.85,
            "bounding_box": [100, 150, 300, 400],
            "estimated_cost": {
                "min": 1200,
                "max": 1500
            },
            "recommendations": [
                "Replace front bumper",
                "Paint matching required",
                "Check impact absorber"
            ]
        }

def analyze_damage(image_path: str) -> Dict[str, Any]:
    """Main function to analyze vehicle damage"""
    analyzer = DamageAnalyzer()
    
    # Ensure image exists
    if not Path(image_path).exists():
        raise FileNotFoundError(f"Image not found: {image_path}")
    
    # Process image and get analysis
    try:
        image_array = analyzer.preprocess_image(image_path)
        results = analyzer.analyze(image_array)
        return results
    except Exception as e:
        raise Exception(f"Error analyzing image: {str(e)}")