backend/requirements.txt
fastapi==0.110.0
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4
pandas==2.2.1
uvicorn==0.27.1
python-jose==3.3.0
python-dotenv==1.0.1